Restructure
-- renamed a few files. Conforming to PurpleAirMODULENAMEHERE.py

-- Update the docs for renamed files

-- added more doc comments
