+++ docs/html/PurpleAirAPI.html
@@ -43,8 +43,8 @@
 <li class="toctree-l1 current"><a class="reference internal" href="modules.html">purple_air_data_logger</a><ul class="current">
 <li class="toctree-l2 current"><a class="current reference internal" href="#">PurpleAirAPI module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="PurpleAirAPIConstants.html">PurpleAirAPIConstants module</a></li>
-<li class="toctree-l2"><a class="reference internal" href="PurpleAirDataLogger.html">PurpleAirDataLogger module</a></li>
-<li class="toctree-l2"><a class="reference internal" href="PurpleAirDataLoggerPSQLStatements.html">PurpleAirDataLoggerPSQLStatements module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="PurpleAirPSQLDataLogger.html">PurpleAirPSQLDataLogger module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="PurpleAirPSQLQueryStatements.html">PurpleAirPSQLQueryStatements module</a></li>
 </ul>
 </li>
 </ul>
@@ -200,7 +200,9 @@ information:
 <dt class="sig sig-object py" id="PurpleAirAPI.debug_log">
 <span class="sig-prename descclassname"><span class="pre">PurpleAirAPI.</span></span><span class="sig-name descname"><span class="pre">debug_log</span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">debug_msg_string</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#PurpleAirAPI.debug_log" title="Permalink to this definition"></a></dt>
 <dd><p>A helper function to print out
-debug messages only if DEBUG is defined</p>
+debug messages only if DEBUG is defined as True in
+‘PurpleAirAPIConstants.py’. Messages will be the color
+red.</p>
 </dd></dl>
 
 </section>
