--- docs/html/_sources/PurpleAirDataLogger.rst.txt
@@ -1,7 +1,0 @@
-PurpleAirDataLogger module
-==========================
-
-.. automodule:: PurpleAirDataLogger
-   :members:
-   :undoc-members:
-   :show-inheritance:
