--- docs/html/_sources/PurpleAirDataLoggerPSQLStatements.rst.txt
@@ -1,7 +1,0 @@
-PurpleAirDataLoggerPSQLStatements module
-========================================
-
-.. automodule:: PurpleAirDataLoggerPSQLStatements
-   :members:
-   :undoc-members:
-   :show-inheritance:
