+++ docs/html/_sources/PurpleAirPSQLDataLogger.rst.txt
@@ -1,0 +1,7 @@
+PurpleAirPSQLDataLogger module
+==============================
+
+.. automodule:: PurpleAirPSQLDataLogger
+   :members:
+   :undoc-members:
+   :show-inheritance:
