+++ docs/html/_sources/PurpleAirPSQLQueryStatements.rst.txt
@@ -1,0 +1,7 @@
+PurpleAirPSQLQueryStatements module
+===================================
+
+.. automodule:: PurpleAirPSQLQueryStatements
+   :members:
+   :undoc-members:
+   :show-inheritance:
