+++ docs/html/index.html
@@ -74,8 +74,8 @@
 <li class="toctree-l1"><a class="reference internal" href="modules.html">purple_air_data_logger</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="PurpleAirAPI.html">PurpleAirAPI module</a></li>
 <li class="toctree-l2"><a class="reference internal" href="PurpleAirAPIConstants.html">PurpleAirAPIConstants module</a></li>
-<li class="toctree-l2"><a class="reference internal" href="PurpleAirDataLogger.html">PurpleAirDataLogger module</a></li>
-<li class="toctree-l2"><a class="reference internal" href="PurpleAirDataLoggerPSQLStatements.html">PurpleAirDataLoggerPSQLStatements module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="PurpleAirPSQLDataLogger.html">PurpleAirPSQLDataLogger module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="PurpleAirPSQLQueryStatements.html">PurpleAirPSQLQueryStatements module</a></li>
 </ul>
 </li>
 </ul>
