+++ docs/html/py-modindex.html
@@ -94,12 +94,12 @@
      <tr>
        <td></td>
        <td>
-       <a href="PurpleAirDataLogger.html#module-PurpleAirDataLogger"><code class="xref">PurpleAirDataLogger</code></a></td><td>
+       <a href="PurpleAirPSQLDataLogger.html#module-PurpleAirPSQLDataLogger"><code class="xref">PurpleAirPSQLDataLogger</code></a></td><td>
        <em></em></td></tr>
      <tr>
        <td></td>
        <td>
-       <a href="PurpleAirDataLoggerPSQLStatements.html#module-PurpleAirDataLoggerPSQLStatements"><code class="xref">PurpleAirDataLoggerPSQLStatements</code></a></td><td>
+       <a href="PurpleAirPSQLQueryStatements.html#module-PurpleAirPSQLQueryStatements"><code class="xref">PurpleAirPSQLQueryStatements</code></a></td><td>
        <em></em></td></tr>
    </table>
