+++ docs/html/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({"docnames": ["PurpleAirAPI", "PurpleAirAPIConstants", "PurpleAirDataLogger", "PurpleAirDataLoggerPSQLStatements", "index", "modules"], "filenames": ["PurpleAirAPI.rst", "PurpleAirAPIConstants.rst", "PurpleAirDataLogger.rst", "PurpleAirDataLoggerPSQLStatements.rst", "index.rst", "modules.rst"], "titles": ["PurpleAirAPI module", "PurpleAirAPIConstants module", "PurpleAirDataLogger module", "PurpleAirDataLoggerPSQLStatements module", "Welcome to PurpleAir Data Logger\u2019s documentation!", "purple_air_data_logger"], "terms": {"copyright": [0, 1, 2, 3], "2022": [0, 1, 2, 3], "carlkid1499": [0, 1, 2, 3], "all": [0, 1, 2, 3], "right": [0, 1, 2, 3], "reserv": [0, 1, 2, 3], "A": [0, 1, 2, 3], "python3": [0, 1], "class": [0, 2], "design": [0, 2], "fetch": 0, "data": [0, 1, 2], "from": [0, 1, 2], "purpl": 0, "air": 0, "s": [0, 2], "new": 0, "api": [0, 1, 2], "http": [0, 1], "purpleair": [0, 1, 2, 3], "com": [0, 1], "welcom": 0, "your_api_read_kei": 0, "base": [0, 2], "object": [0, 2], "The": [0, 2], "send": [0, 2], "onli": [0, 2], "read": [0, 2], "request": [0, 2], "properti": 0, "get_api_key_last_check": 0, "method": [0, 2], "return": [0, 2], "timestamp": 0, "when": 0, "kei": [0, 1, 2, 3], "wa": 0, "last": 0, "check": [0, 2], "get_api_key_typ": 0, "version": 0, "being": 0, "us": [0, 1, 2, 3], "get_api_vers": 0, "recheck_api_kei": 0, "recheck": 0, "provid": 0, "request_multiple_sensors_data": [0, 2], "field": [0, 1, 2], "location_typ": [0, 1, 2, 3], "none": [0, 2], "read_kei": [0, 2], "show_onli": [0, 2], "modified_sinc": [0, 2], "max_ag": [0, 2], "nwlng": [0, 2], "nwlat": [0, 2], "selng": [0, 2], "selat": [0, 2], "retriev": [0, 2], "sensor": [0, 1, 2], "multipl": [0, 2], "Will": 0, "respons": 0, "payload": 0, "python": [0, 2], "dictionari": [0, 2], "paramet": [0, 2], "str": [0, 2], "specifi": 0, "which": 0, "includ": 0, "It": 0, "comma": [0, 2], "separ": 0, "list": 0, "one": 0, "more": [0, 2], "follow": 0, "station": 0, "inform": 0, "statu": 0, "name": [0, 1, 2, 3], "icon": [0, 1, 3], "model": [0, 1, 3], "hardwar": [0, 1, 3], "privat": [0, 1, 2, 3], "latitud": [0, 1, 3], "longitud": [0, 1, 3], "altitud": [0, 1, 3], "position_r": [0, 1, 3], "led_bright": [0, 1, 3], "firmware_vers": [0, 1, 3], "firmware_upgrad": [0, 1, 3], "rssi": [0, 1, 3], "uptim": [0, 1, 3], "pa_lat": [0, 1, 3], "memori": [0, 1, 3], "last_seen": [0, 1, 3], "last_modifi": [0, 1, 3], "date_cr": [0, 1, 3], "channel_st": [0, 1, 3], "channel_flag": [0, 1, 3], "channel_flags_manu": [0, 1, 3], "channel_flags_auto": [0, 1, 3], "confid": [0, 1, 3], "confidence_manu": [0, 1, 3], "confidence_auto": [0, 1, 3], "environment": 0, "humid": [0, 1, 3], "humidity_a": [0, 1, 3], "humidity_b": [0, 1, 3], "temperatur": [0, 1, 3], "temperature_a": [0, 1, 3], "temperature_b": [0, 1, 3], "pressur": [0, 1, 3], "pressure_a": [0, 1, 3], "pressure_b": [0, 1, 3], "miscellan": 0, "voc": [0, 1, 3], "voc_a": [0, 1, 3], "voc_b": [0, 1, 3], "ozone1": [0, 1, 3], "analog_input": [0, 1, 3], "pm1": [0, 1, 3], "0": [0, 1, 3], "0_a": [0, 1], "0_b": [0, 1], "0_atm": [0, 1], "0_atm_a": [0, 1], "0_atm_b": [0, 1], "0_cf_1": [0, 1], "0_cf_1_a": [0, 1], "0_cf_1_b": [0, 1], "pm2": [0, 1, 3], "5": [0, 1, 3], "5_alt": [0, 1], "5_alt_a": [0, 1], "5_alt_b": [0, 1], "5_a": [0, 1], "5_b": [0, 1], "5_atm": [0, 1], "5_atm_a": [0, 1], "5_atm_b": [0, 1], "5_cf_1": [0, 1], "5_cf_1_a": [0, 1], "5_cf_1_b": [0, 1], "pseudo": 0, "simpl": 0, "run": [0, 2], "averag": 0, "5_10minut": [0, 1], "5_10minute_a": [0, 1], "5_10minute_b": [0, 1], "5_30minut": [0, 1], "5_30minute_a": [0, 1], "5_30minute_b": [0, 1], "5_60minut": [0, 1], "5_60minute_a": [0, 1], "5_60minute_b": [0, 1], "5_6hour": [0, 1], "5_6hour_a": [0, 1], "5_6hour_b": [0, 1], "5_24hour": [0, 1], "5_24hour_a": [0, 1], "5_24hour_b": [0, 1], "5_1week": [0, 1], "5_1week_a": [0, 1], "5_1week_b": [0, 1], "pm10": [0, 1, 3], "visibl": 0, "scattering_coeffici": 0, "scattering_coefficient_a": 0, "scattering_coefficient_b": 0, "deciview": 0, "deciviews_a": 0, "deciviews_b": 0, "visual_rang": 0, "visual_range_a": 0, "visual_range_b": 0, "particl": 0, "count": 0, "3_um_count": [0, 1], "3_um_count_a": [0, 1], "3_um_count_b": [0, 1], "5_um_count": [0, 1], "5_um_count_a": [0, 1], "5_um_count_b": [0, 1], "1": [0, 1, 2], "0_um_count": [0, 1], "0_um_count_a": [0, 1], "0_um_count_b": [0, 1], "2": [0, 1], "10": [0, 1, 2], "thingspeak": 0, "primary_id_a": [0, 1, 3], "primary_key_a": [0, 1, 3], "secondary_id_a": [0, 1, 3], "secondary_key_a": [0, 1, 3], "primary_id_b": [0, 1, 3], "primary_key_b": [0, 1, 3], "secondary_id_b": [0, 1, 3], "secondary_key_b": [0, 1, 3], "option": 0, "int": [0, 2, 3], "possibl": 0, "valu": [0, 3], "ar": [0, 1, 2], "outsid": 0, "insid": 0, "requir": 0, "devic": 0, "api_kei": 0, "each": 0, "ha": 0, "its": 0, "own": 0, "submit": 0, "them": 0, "charact": 0, "exampl": 0, "two": 0, "three": 0, "sensor_index": [0, 2, 3], "result": 0, "limit": [0, 2], "thi": [0, 2, 3], "caus": 0, "modifi": 0, "after": 0, "time": 0, "stamp": 0, "time_stamp": 0, "previou": 0, "call": 0, "recommend": 0, "those": 0, "sinc": [0, 3], "match": [0, 3], "ani": 0, "filter": 0, "updat": [0, 2], "within": 0, "number": [0, 2, 3], "second": [0, 2], "ag": 0, "default": 0, "604800": 0, "north": 0, "west": 0, "bound": 0, "box": 0, "specif": 0, "geograph": 0, "area": 0, "defin": [0, 3], "point": 0, "south": 0, "east": 0, "contain": [0, 2, 3], "request_sensor_data": 0, "found": 0, "json": [0, 1], "refer": 0, "document": [0, 1, 3], "get": [0, 1, 3], "debug_log": 0, "debug_msg_str": 0, "helper": 0, "function": 0, "print": 0, "out": 0, "debug": 0, "messag": 0, "purpleairapi": [2, 4, 5], "For": [2, 3], "best": 2, "practic": 2, "individu": 2, "less": 2, "than": 2, "everi": 2, "30": 2, "As": 2, "courtesi": 2, "we": [2, 3], "ask": 2, "you": 2, "onc": 2, "minut": 2, "assum": 2, "obtain": 2, "If": 2, "pleas": 2, "singl": 2, "rather": 2, "success": 2, "purpleairapireadkei": 2, "psql_db_conn": 2, "logger": 2, "now": 2, "ingest": 2, "timescaledb": 2, "postgresql": 2, "databas": 2, "Then": 2, "grafana": 2, "visual": 2, "said": 2, "get_accepted_field_names_list": [], "accept": [], "string": [2, 3], "get_multiple_sensors_data": 2, "same": [2, 3], "get_sensor_data": 2, "valid": 2, "index": [2, 4], "delmit": 2, "run_loop_for_storing_multiple_sensors_data": 2, "loop": 2, "insert": [2, 3], "db": 2, "run_loop_for_storing_single_sensor_data": 2, "store_multiple_sensors_data": 2, "store_sensor_data": 2, "single_sensor_data_dict": 2, "dict": [1, 2], "doesn": 2, "t": [2, 3], "support": 2, "certain": 2, "make": 2, "sure": 2, "null": 2, "part": 2, "doe": 2, "type": 2, "error": 2, "That": 2, "upto": 2, "caller": 2, "file": 3, "psql": 3, "statement": 3, "constant": [1, 3], "purple_air_data_logg": 4, "modul": [4, 5], "purpleairdatalogg": [4, 5], "purpleairdataloggerpsqlstat": [4, 5], "search": 4, "page": 4, "create_environmental_fields_t": 3, "n": 3, "creat": 3, "tabl": 3, "IF": 3, "NOT": 3, "exist": 3, "environmental_field": 3, "data_time_stamp": 3, "timestamptz": 3, "primari": 3, "float": 3, "statemnt": 3, "create_miscellaneous_field": 3, "miscellaneous_field": 3, "create_particle_count_field": 3, "particle_count_field": 3, "um_count_0_3": 3, "um_count_a_0_3": 3, "um_count_b_0_3": 3, "um_count_0_5": 3, "um_count_a_0_5": 3, "um_count_b_0_5": 3, "um_count_1_0": 3, "um_count_a_1_0": 3, "um_count_b_1_0": 3, "um_count_2_5": 3, "um_count_a_2_5": 3, "um_count_b_2_5": 3, "um_count_5_0": 3, "um_count_a_5_0": 3, "um_count_b_5_0": 3, "um_count_10_0": 3, "um_count_a_10_0": 3, "um_count_b_10_0": 3, "create_pm10_0_field": 3, "pm10_0_field": 3, "pm10_0": 3, "pm10_0_a": 3, "pm10_0_b": 3, "pm10_0_atm": 3, "pm10_0_atm_a": 3, "pm10_0_atm_b": 3, "pm10_0_cf_1": 3, "pm10_0_cf_1_a": 3, "pm10_0_cf_1_b": 3, "create_pm1_0_field": 3, "pm1_0_field": 3, "pm1_0": 3, "pm1_0_a": 3, "pm1_0_b": 3, "pm1_0_atm": 3, "pm1_0_atm_a": 3, "pm1_0_atm_b": 3, "pm1_0_cf_1": 3, "pm1_0_cf_1_a": 3, "pm1_0_cf_1_b": 3, "create_pm2_5_field": 3, "pm2_5_field": 3, "pm2_5_alt": 3, "pm2_5_alt_a": 3, "pm2_5_alt_b": 3, "pm2_5": 3, "pm2_5_a": 3, "pm2_5_b": 3, "pm2_5_atm": 3, "pm2_5_atm_a": 3, "pm2_5_atm_b": 3, "pm2_5_cf_1": 3, "pm2_5_cf_1_a": 3, "pm2_5_cf_1_b": 3, "create_pm2_5_pseudo_average_field": 3, "pm2_5_pseudo_average_field": 3, "pm2_5_10minut": 3, "pm2_5_10minute_a": 3, "pm2_5_10minute_b": 3, "pm2_5_30minut": 3, "pm2_5_30minute_a": 3, "pm2_5_30minute_b": 3, "pm2_5_60minut": 3, "pm2_5_60minute_a": 3, "pm2_5_60minute_b": 3, "pm2_5_6hour": 3, "pm2_5_6hour_a": 3, "pm2_5_6hour_b": 3, "pm2_5_24hour": 3, "pm2_5_24hour_a": 3, "pm2_5_24hour_b": 3, "pm2_5_1week": 3, "pm2_5_1week_a": 3, "pm2_5_1week_b": 3, "create_station_information_and_status_fields_t": 3, "station_information_and_status_field": 3, "text": 3, "create_thingspeak_field": 3, "thingspeak_field": 3, "gener": 3, "queri": 3, "simplic": 3, "our": 3, "what": 3, "sai": 3, "do": 3, "column": 3, "can": 3, "have": 3, "decim": 3, "variabl": 3, "instead": 3, "note": 3, "start": 3, "so": 3, "0_3_um_count": 3, "becom": 3, "TO": 3, "self": 3, "mai": 3, "end": 3, "up": 3, "rid": 3, "OF": 3, "i": 3, "see": 3, "NO": 3, "FOR": 3, "IT": 3, "psql_insert_statement_environmental_field": 3, "INTO": 3, "cast": 3, "AS": 3, "psql_insert_statement_miscellaneous_field": 3, "psql_insert_statement_station_information_and_status_field": 3, "psql_insert_statement_particle_count_field": 3, "psql_insert_statement_pm10_0_field": 3, "psql_insert_statement_pm1_0_field": 3, "psql_insert_statement_pm2_5_field": 3, "psql_insert_statement_pm2_5_pseudo_average_field": 3, "psql_insert_statement_thingspeak_field": 3, "purpleairapiconst": [4, 5], "accepted_field_names_list": 1, "store": 1, "access": 1, "These": 1, "deriv": 1}, "objects": {"": [[0, 0, 0, "-", "PurpleAirAPI"], [1, 0, 0, "-", "PurpleAirAPIConstants"], [2, 0, 0, "-", "PurpleAirDataLogger"], [3, 0, 0, "-", "PurpleAirDataLoggerPSQLStatements"]], "PurpleAirAPI": [[0, 1, 1, "", "PurpleAirAPI"], [0, 4, 1, "", "debug_log"]], "PurpleAirAPI.PurpleAirAPI": [[0, 2, 1, "", "get_api_key_last_checked"], [0, 2, 1, "", "get_api_key_type"], [0, 2, 1, "", "get_api_version"], [0, 3, 1, "", "recheck_api_key"], [0, 3, 1, "", "request_multiple_sensors_data"], [0, 3, 1, "", "request_sensor_data"]], "PurpleAirAPIConstants": [[1, 5, 1, "", "ACCEPTED_FIELD_NAMES_LIST"]], "PurpleAirDataLogger": [[2, 1, 1, "", "PurpleAirDataLogger"]], "PurpleAirDataLogger.PurpleAirDataLogger": [[2, 3, 1, "", "get_multiple_sensors_data"], [2, 3, 1, "", "get_sensor_data"], [2, 3, 1, "", "run_loop_for_storing_multiple_sensors_data"], [2, 3, 1, "", "run_loop_for_storing_single_sensor_data"], [2, 3, 1, "", "store_multiple_sensors_data"], [2, 3, 1, "", "store_sensor_data"]], "PurpleAirDataLoggerPSQLStatements": [[3, 5, 1, "", "CREATE_ENVIRONMENTAL_FIELDS_TABLE"], [3, 5, 1, "", "CREATE_MISCELLANEOUS_FIELDS"], [3, 5, 1, "", "CREATE_PARTICLE_COUNT_FIELDS"], [3, 5, 1, "", "CREATE_PM10_0_FIELDS"], [3, 5, 1, "", "CREATE_PM1_0_FIELDS"], [3, 5, 1, "", "CREATE_PM2_5_FIELDS"], [3, 5, 1, "", "CREATE_PM2_5_PSEUDO_AVERAGE_FIELDS"], [3, 5, 1, "", "CREATE_STATION_INFORMATION_AND_STATUS_FIELDS_TABLE"], [3, 5, 1, "", "CREATE_THINGSPEAK_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_ENVIRONMENTAL_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_MISCELLANEOUS_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PARTICLE_COUNT_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PM10_0_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PM1_0_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PM2_5_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PM2_5_PSEUDO_AVERAGE_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_STATION_INFORMATION_AND_STATUS_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_THINGSPEAK_FIELDS"]]}, "objtypes": {"0": "py:module", "1": "py:class", "2": "py:property", "3": "py:method", "4": "py:function", "5": "py:data"}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "class", "Python class"], "2": ["py", "property", "Python property"], "3": ["py", "method", "Python method"], "4": ["py", "function", "Python function"], "5": ["py", "data", "Python data"]}, "titleterms": {"purpleairapi": 0, "modul": [0, 1, 2, 3], "purpleairdatalogg": 2, "purpleairdataloggerpsqlstat": 3, "welcom": 4, "purpleair": 4, "data": 4, "logger": 4, "s": 4, "document": 4, "content": 4, "indic": 4, "tabl": 4, "purple_air_data_logg": 5, "purpleairapiconst": 1}, "envversion": {"sphinx.domains.c": 2, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 6, "sphinx.domains.index": 1, "sphinx.domains.javascript": 2, "sphinx.domains.math": 2, "sphinx.domains.python": 3, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx": 56}})
+Search.setIndex({"docnames": ["PurpleAirAPI", "PurpleAirAPIConstants", "PurpleAirPSQLDataLogger", "PurpleAirPSQLQueryStatements", "index", "modules"], "filenames": ["PurpleAirAPI.rst", "PurpleAirAPIConstants.rst", "PurpleAirPSQLDataLogger.rst", "PurpleAirPSQLQueryStatements.rst", "index.rst", "modules.rst"], "titles": ["PurpleAirAPI module", "PurpleAirAPIConstants module", "PurpleAirPSQLDataLogger module", "PurpleAirPSQLQueryStatements module", "Welcome to PurpleAir Data Logger\u2019s documentation!", "purple_air_data_logger"], "terms": {"copyright": [0, 1, 2, 3], "2022": [0, 1, 2, 3], "carlkid1499": [0, 1, 2, 3], "all": [0, 1, 2, 3], "right": [0, 1, 2, 3], "reserv": [0, 1, 2, 3], "A": [0, 1, 2, 3], "python3": 0, "class": [0, 2], "design": [0, 2], "fetch": 0, "data": [0, 1, 2], "from": [0, 1, 2], "purpl": 0, "air": 0, "s": [0, 2], "new": 0, "api": [0, 1, 2], "http": [0, 1], "purpleair": [0, 1, 2, 3], "com": [0, 1], "welcom": 0, "your_api_read_kei": 0, "base": [0, 2], "object": [0, 2], "The": [0, 2], "send": [0, 2], "onli": [0, 2], "read": [0, 2], "request": [0, 2], "properti": 0, "get_api_key_last_check": 0, "method": [0, 2], "return": [0, 2], "timestamp": 0, "when": 0, "kei": [0, 1, 2, 3], "wa": 0, "last": 0, "check": [0, 2], "get_api_key_typ": 0, "version": 0, "being": 0, "us": [0, 1, 2, 3], "get_api_vers": 0, "recheck_api_kei": 0, "recheck": 0, "provid": 0, "request_multiple_sensors_data": [0, 2], "field": [0, 1, 2], "location_typ": [0, 1, 2, 3], "none": [0, 2], "read_kei": [0, 2], "show_onli": [0, 2], "modified_sinc": [0, 2], "max_ag": [0, 2], "nwlng": [0, 2], "nwlat": [0, 2], "selng": [0, 2], "selat": [0, 2], "retriev": [0, 2], "sensor": [0, 1, 2], "multipl": [0, 2], "Will": 0, "respons": 0, "payload": 0, "python": [0, 1, 2], "dictionari": [0, 2], "paramet": [0, 2], "str": [0, 2], "specifi": 0, "which": 0, "includ": 0, "It": 0, "comma": [0, 2], "separ": 0, "list": 0, "one": 0, "more": [0, 2], "follow": 0, "station": 0, "inform": 0, "statu": 0, "name": [0, 1, 2, 3], "icon": [0, 1, 3], "model": [0, 1, 3], "hardwar": [0, 1, 3], "privat": [0, 1, 2, 3], "latitud": [0, 1, 3], "longitud": [0, 1, 3], "altitud": [0, 1, 3], "position_r": [0, 1, 3], "led_bright": [0, 1, 3], "firmware_vers": [0, 1, 3], "firmware_upgrad": [0, 1, 3], "rssi": [0, 1, 3], "uptim": [0, 1, 3], "pa_lat": [0, 1, 3], "memori": [0, 1, 3], "last_seen": [0, 1, 3], "last_modifi": [0, 1, 3], "date_cr": [0, 1, 3], "channel_st": [0, 1, 3], "channel_flag": [0, 1, 3], "channel_flags_manu": [0, 1, 3], "channel_flags_auto": [0, 1, 3], "confid": [0, 1, 3], "confidence_manu": [0, 1, 3], "confidence_auto": [0, 1, 3], "environment": 0, "humid": [0, 1, 3], "humidity_a": [0, 1, 3], "humidity_b": [0, 1, 3], "temperatur": [0, 1, 3], "temperature_a": [0, 1, 3], "temperature_b": [0, 1, 3], "pressur": [0, 1, 3], "pressure_a": [0, 1, 3], "pressure_b": [0, 1, 3], "miscellan": 0, "voc": [0, 1, 3], "voc_a": [0, 1, 3], "voc_b": [0, 1, 3], "ozone1": [0, 1, 3], "analog_input": [0, 1, 3], "pm1": [0, 1, 3], "0": [0, 1, 3], "0_a": [0, 1], "0_b": [0, 1], "0_atm": [0, 1], "0_atm_a": [0, 1], "0_atm_b": [0, 1], "0_cf_1": [0, 1], "0_cf_1_a": [0, 1], "0_cf_1_b": [0, 1], "pm2": [0, 1, 3], "5": [0, 1, 3], "5_alt": [0, 1], "5_alt_a": [0, 1], "5_alt_b": [0, 1], "5_a": [0, 1], "5_b": [0, 1], "5_atm": [0, 1], "5_atm_a": [0, 1], "5_atm_b": [0, 1], "5_cf_1": [0, 1], "5_cf_1_a": [0, 1], "5_cf_1_b": [0, 1], "pseudo": 0, "simpl": 0, "run": [0, 2], "averag": 0, "5_10minut": [0, 1], "5_10minute_a": [0, 1], "5_10minute_b": [0, 1], "5_30minut": [0, 1], "5_30minute_a": [0, 1], "5_30minute_b": [0, 1], "5_60minut": [0, 1], "5_60minute_a": [0, 1], "5_60minute_b": [0, 1], "5_6hour": [0, 1], "5_6hour_a": [0, 1], "5_6hour_b": [0, 1], "5_24hour": [0, 1], "5_24hour_a": [0, 1], "5_24hour_b": [0, 1], "5_1week": [0, 1], "5_1week_a": [0, 1], "5_1week_b": [0, 1], "pm10": [0, 1, 3], "visibl": 0, "scattering_coeffici": 0, "scattering_coefficient_a": 0, "scattering_coefficient_b": 0, "deciview": 0, "deciviews_a": 0, "deciviews_b": 0, "visual_rang": 0, "visual_range_a": 0, "visual_range_b": 0, "particl": 0, "count": 0, "3_um_count": [0, 1], "3_um_count_a": [0, 1], "3_um_count_b": [0, 1], "5_um_count": [0, 1], "5_um_count_a": [0, 1], "5_um_count_b": [0, 1], "1": [0, 1, 2], "0_um_count": [0, 1], "0_um_count_a": [0, 1], "0_um_count_b": [0, 1], "2": [0, 1], "10": [0, 1, 2], "thingspeak": 0, "primary_id_a": [0, 1, 3], "primary_key_a": [0, 1, 3], "secondary_id_a": [0, 1, 3], "secondary_key_a": [0, 1, 3], "primary_id_b": [0, 1, 3], "primary_key_b": [0, 1, 3], "secondary_id_b": [0, 1, 3], "secondary_key_b": [0, 1, 3], "option": 0, "int": [0, 2, 3], "possibl": 0, "valu": [0, 3], "ar": [0, 1, 2], "outsid": 0, "insid": 0, "requir": 0, "devic": 0, "api_kei": 0, "each": 0, "ha": 0, "its": 0, "own": 0, "submit": 0, "them": 0, "charact": 0, "exampl": 0, "two": 0, "three": 0, "sensor_index": [0, 2, 3], "result": 0, "limit": [0, 2], "thi": [0, 2, 3], "caus": 0, "modifi": 0, "after": 0, "time": 0, "stamp": 0, "time_stamp": 0, "previou": 0, "call": 0, "recommend": 0, "those": 0, "sinc": [0, 3], "match": [0, 3], "ani": 0, "filter": 0, "updat": [0, 2], "within": 0, "number": [0, 2, 3], "second": [0, 2], "ag": 0, "default": 0, "604800": 0, "north": 0, "west": 0, "bound": 0, "box": 0, "specif": 0, "geograph": 0, "area": 0, "defin": [0, 3], "point": 0, "south": 0, "east": 0, "contain": [0, 2, 3], "request_sensor_data": 0, "found": 0, "json": [0, 1], "refer": 0, "document": [0, 1, 3], "get": [0, 1, 3], "debug_log": 0, "debug_msg_str": 0, "helper": 0, "function": 0, "print": 0, "out": 0, "debug": [0, 1], "messag": 0, "true": [0, 1], "purpleairapiconst": [0, 4, 5], "py": [0, 1], "color": 0, "red": 0, "constant": [1, 3], "purpleairapi": [1, 2, 4, 5], "accepted_field_names_list": 1, "store": 1, "dict": [1, 2], "access": 1, "These": 1, "deriv": 1, "print_debug_msg": 1, "see": [1, 3], "statement": [1, 3], "enabl": 1, "For": [2, 3], "best": 2, "practic": 2, "individu": 2, "less": 2, "than": 2, "everi": 2, "30": 2, "As": 2, "courtesi": 2, "we": [2, 3], "ask": 2, "you": 2, "onc": 2, "minut": 2, "assum": 2, "obtain": 2, "If": 2, "pleas": 2, "singl": 2, "rather": 2, "success": 2, "purpleairdatalogg": 2, "purpleairapireadkei": 2, "psql_db_conn": 2, "logger": 2, "now": 2, "ingest": 2, "timescaledb": 2, "postgresql": 2, "databas": 2, "Then": 2, "grafana": 2, "visual": 2, "said": 2, "get_multiple_sensors_data": 2, "same": [2, 3], "get_sensor_data": 2, "valid": 2, "index": [2, 4], "delmit": 2, "string": [2, 3], "run_loop_for_storing_multiple_sensors_data": 2, "loop": 2, "insert": [2, 3], "db": 2, "run_loop_for_storing_single_sensor_data": 2, "store_multiple_sensors_data": 2, "store_sensor_data": 2, "single_sensor_data_dict": 2, "doesn": 2, "t": [2, 3], "support": 2, "certain": 2, "make": 2, "sure": 2, "null": 2, "part": 2, "doe": 2, "type": 2, "error": 2, "That": 2, "upto": 2, "caller": 2, "file": 3, "psql": 3, "gener": 3, "queri": 3, "simplic": 3, "our": 3, "tabl": 3, "what": 3, "sai": 3, "do": 3, "column": 3, "create_environmental_fields_t": 3, "n": 3, "creat": 3, "IF": 3, "NOT": 3, "exist": 3, "environmental_field": 3, "data_time_stamp": 3, "timestamptz": 3, "primari": 3, "float": 3, "statemnt": 3, "create_miscellaneous_field": 3, "miscellaneous_field": 3, "create_particle_count_field": 3, "particle_count_field": 3, "um_count_0_3": 3, "um_count_a_0_3": 3, "um_count_b_0_3": 3, "um_count_0_5": 3, "um_count_a_0_5": 3, "um_count_b_0_5": 3, "um_count_1_0": 3, "um_count_a_1_0": 3, "um_count_b_1_0": 3, "um_count_2_5": 3, "um_count_a_2_5": 3, "um_count_b_2_5": 3, "um_count_5_0": 3, "um_count_a_5_0": 3, "um_count_b_5_0": 3, "um_count_10_0": 3, "um_count_a_10_0": 3, "um_count_b_10_0": 3, "note": 3, "can": 3, "start": 3, "so": 3, "0_3_um_count": 3, "becom": 3, "create_pm10_0_field": 3, "pm10_0_field": 3, "pm10_0": 3, "pm10_0_a": 3, "pm10_0_b": 3, "pm10_0_atm": 3, "pm10_0_atm_a": 3, "pm10_0_atm_b": 3, "pm10_0_cf_1": 3, "pm10_0_cf_1_a": 3, "pm10_0_cf_1_b": 3, "have": 3, "decim": 3, "variabl": 3, "instead": 3, "create_pm1_0_field": 3, "pm1_0_field": 3, "pm1_0": 3, "pm1_0_a": 3, "pm1_0_b": 3, "pm1_0_atm": 3, "pm1_0_atm_a": 3, "pm1_0_atm_b": 3, "pm1_0_cf_1": 3, "pm1_0_cf_1_a": 3, "pm1_0_cf_1_b": 3, "create_pm2_5_field": 3, "pm2_5_field": 3, "pm2_5_alt": 3, "pm2_5_alt_a": 3, "pm2_5_alt_b": 3, "pm2_5": 3, "pm2_5_a": 3, "pm2_5_b": 3, "pm2_5_atm": 3, "pm2_5_atm_a": 3, "pm2_5_atm_b": 3, "pm2_5_cf_1": 3, "pm2_5_cf_1_a": 3, "pm2_5_cf_1_b": 3, "create_pm2_5_pseudo_average_field": 3, "pm2_5_pseudo_average_field": 3, "pm2_5_10minut": 3, "pm2_5_10minute_a": 3, "pm2_5_10minute_b": 3, "pm2_5_30minut": 3, "pm2_5_30minute_a": 3, "pm2_5_30minute_b": 3, "pm2_5_60minut": 3, "pm2_5_60minute_a": 3, "pm2_5_60minute_b": 3, "pm2_5_6hour": 3, "pm2_5_6hour_a": 3, "pm2_5_6hour_b": 3, "pm2_5_24hour": 3, "pm2_5_24hour_a": 3, "pm2_5_24hour_b": 3, "pm2_5_1week": 3, "pm2_5_1week_a": 3, "pm2_5_1week_b": 3, "create_station_information_and_status_fields_t": 3, "station_information_and_status_field": 3, "text": 3, "create_thingspeak_field": 3, "thingspeak_field": 3, "TO": 3, "self": 3, "mai": 3, "end": 3, "up": 3, "rid": 3, "OF": 3, "i": 3, "NO": 3, "FOR": 3, "IT": 3, "psql_insert_statement_environmental_field": 3, "INTO": 3, "cast": 3, "AS": 3, "psql_insert_statement_miscellaneous_field": 3, "psql_insert_statement_particle_count_field": 3, "psql_insert_statement_pm10_0_field": 3, "psql_insert_statement_pm1_0_field": 3, "psql_insert_statement_pm2_5_field": 3, "psql_insert_statement_pm2_5_pseudo_average_field": 3, "psql_insert_statement_station_information_and_status_field": 3, "psql_insert_statement_thingspeak_field": 3, "purple_air_data_logg": 4, "modul": [4, 5], "purpleairpsqldatalogg": [4, 5], "purpleairpsqlquerystat": [4, 5], "search": 4, "page": 4}, "objects": {"": [[0, 0, 0, "-", "PurpleAirAPI"], [1, 0, 0, "-", "PurpleAirAPIConstants"], [2, 0, 0, "-", "PurpleAirPSQLDataLogger"], [3, 0, 0, "-", "PurpleAirPSQLQueryStatements"]], "PurpleAirAPI": [[0, 1, 1, "", "PurpleAirAPI"], [0, 4, 1, "", "debug_log"]], "PurpleAirAPI.PurpleAirAPI": [[0, 2, 1, "", "get_api_key_last_checked"], [0, 2, 1, "", "get_api_key_type"], [0, 2, 1, "", "get_api_version"], [0, 3, 1, "", "recheck_api_key"], [0, 3, 1, "", "request_multiple_sensors_data"], [0, 3, 1, "", "request_sensor_data"]], "PurpleAirAPIConstants": [[1, 5, 1, "", "ACCEPTED_FIELD_NAMES_LIST"], [1, 5, 1, "", "PRINT_DEBUG_MSGS"]], "PurpleAirPSQLDataLogger": [[2, 1, 1, "", "PurpleAirDataLogger"]], "PurpleAirPSQLDataLogger.PurpleAirDataLogger": [[2, 3, 1, "", "get_multiple_sensors_data"], [2, 3, 1, "", "get_sensor_data"], [2, 3, 1, "", "run_loop_for_storing_multiple_sensors_data"], [2, 3, 1, "", "run_loop_for_storing_single_sensor_data"], [2, 3, 1, "", "store_multiple_sensors_data"], [2, 3, 1, "", "store_sensor_data"]], "PurpleAirPSQLQueryStatements": [[3, 5, 1, "", "CREATE_ENVIRONMENTAL_FIELDS_TABLE"], [3, 5, 1, "", "CREATE_MISCELLANEOUS_FIELDS"], [3, 5, 1, "", "CREATE_PARTICLE_COUNT_FIELDS"], [3, 5, 1, "", "CREATE_PM10_0_FIELDS"], [3, 5, 1, "", "CREATE_PM1_0_FIELDS"], [3, 5, 1, "", "CREATE_PM2_5_FIELDS"], [3, 5, 1, "", "CREATE_PM2_5_PSEUDO_AVERAGE_FIELDS"], [3, 5, 1, "", "CREATE_STATION_INFORMATION_AND_STATUS_FIELDS_TABLE"], [3, 5, 1, "", "CREATE_THINGSPEAK_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_ENVIRONMENTAL_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_MISCELLANEOUS_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PARTICLE_COUNT_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PM10_0_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PM1_0_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PM2_5_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_PM2_5_PSEUDO_AVERAGE_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_STATION_INFORMATION_AND_STATUS_FIELDS"], [3, 5, 1, "", "PSQL_INSERT_STATEMENT_THINGSPEAK_FIELDS"]]}, "objtypes": {"0": "py:module", "1": "py:class", "2": "py:property", "3": "py:method", "4": "py:function", "5": "py:data"}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "class", "Python class"], "2": ["py", "property", "Python property"], "3": ["py", "method", "Python method"], "4": ["py", "function", "Python function"], "5": ["py", "data", "Python data"]}, "titleterms": {"purpleairapi": 0, "modul": [0, 1, 2, 3], "purpleairapiconst": 1, "purpleairpsqldatalogg": 2, "purpleairpsqlquerystat": 3, "welcom": 4, "purpleair": 4, "data": 4, "logger": 4, "s": 4, "document": 4, "content": 4, "indic": 4, "tabl": 4, "purple_air_data_logg": 5}, "envversion": {"sphinx.domains.c": 2, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 6, "sphinx.domains.index": 1, "sphinx.domains.javascript": 2, "sphinx.domains.math": 2, "sphinx.domains.python": 3, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx": 56}})
