--- docs/index.html
@@ -1,1 +1,0 @@
-<meta http-equiv="refresh" content="0; url=./html/index.html" />
