--- sphinx_docs_build/source/PurpleAirDataLogger.rst
@@ -1,7 +1,0 @@
-PurpleAirDataLogger module
-==========================
-
-.. automodule:: PurpleAirDataLogger
-   :members:
-   :undoc-members:
-   :show-inheritance:
