--- sphinx_docs_build/source/PurpleAirDataLoggerPSQLStatements.rst
@@ -1,7 +1,0 @@
-PurpleAirDataLoggerPSQLStatements module
-========================================
-
-.. automodule:: PurpleAirDataLoggerPSQLStatements
-   :members:
-   :undoc-members:
-   :show-inheritance:
