+++ sphinx_docs_build/source/PurpleAirPSQLDataLogger.rst
@@ -1,0 +1,7 @@
+PurpleAirPSQLDataLogger module
+==============================
+
+.. automodule:: PurpleAirPSQLDataLogger
+   :members:
+   :undoc-members:
+   :show-inheritance:
