+++ sphinx_docs_build/source/PurpleAirPSQLQueryStatements.rst
@@ -1,0 +1,7 @@
+PurpleAirPSQLQueryStatements module
+===================================
+
+.. automodule:: PurpleAirPSQLQueryStatements
+   :members:
+   :undoc-members:
+   :show-inheritance:
